исправления

+++ js/actions.js
@@ -2,12 +2,12 @@
 deleteObject = document.getElementById("deleteObject");
 
 deleteObject.onclick = function () {
-    if (currentObject != null) {
+    if (currentObject != null && !isSomePointSelected) {
         deleteFunc();
     }
 }
 document.addEventListener('keydown', function (event) {
-    if (event.code == 'Delete' && currentObject != null) {
+    if (event.code == 'Delete' && currentObject != null && !isSomePointSelected) {
         deleteFunc();
     }
 });
@@ -83,54 +83,18 @@ document.addEventListener('keydown', function (event) {
     }
 });
 
-//UNDO 
-undo = document.getElementById("undo"); 
-undo.onclick = function () {
-    alert(curindex);
-    for (var i = 0; i <= indexes[curindex]; i++) {
-        let last = svgPanel.lastElementChild;
-            if (last.id != "svg_grid") {
-            svgPanel.removeChild(last);
-        /*let first;
-        if (secondStack.length != 0) {
-            first = secondStack.pop();
-        } 
-        if (first != null) {
-            svgPanel.appendChild(first);
-            stack.push(first);
-            flag = true;
-        } */ 
-            stack.push(last);
-        } 
-    }
-    curindex--;
-}
-
-//REDO 
-redo = document.getElementById("redo"); 
-redo.onclick = function () {
-    let last;
-        last = stack.pop();
-        svgPanel.appendChild(last);
+function centralLocation(width, height) {
+    svgPanel.setAttribute('style', 'top: 50%; left: 50%; transform: translate(-50%, -50%);');
+    if (scrollPanel.clientWidth - 15 < width)
+        svgPanel.setAttribute('style', 'transform: translate(0, -50%); left: 15px;');
+    if (scrollPanel.clientHeight - 15 < height)
+        svgPanel.setAttribute('style', 'transform: translate(-50%, 0); top: 15px;');
+    if (scrollPanel.clientWidth - 15 < width && scrollPanel.clientHeight - 15 < height)
+        svgPanel.style.transform = "translate(0, 0)";
 }
 
 //CREATE 
 create = document.getElementById("create");
-
-function deleteChild(node, parent) {
-    for (var i = 0; i < node.childNodes.length;)
-        deleteChild(node.childNodes[i], node);
-
-    if (node.childNodes.length == 0) {
-        parent.removeChild(node);
-        return;
-    }
-}
-
-function deleteAllChildren(node) {
-    for (var i = 2; i < node.childNodes.length;)
-        deleteChild(node.childNodes[i], node);
-}
 
 create.onclick = function () {
     let width = prompt('Введите ширину нового холста в пикселях:', 512);
@@ -141,8 +105,12 @@ create.onclick = function () {
         return;
     }
 
-    deleteAllChildren(svgPanel);
+    for (var i = 2; i < svgPanel.childNodes.length;)
+        svgPanel.removeChild(svgPanel.childNodes[i]);
+    for (var i = 0; i < layersPanel.childNodes.length;)
+        layersPanel.removeChild(layersPanel.childNodes[i]);
 
+    centralLocation(width, height);
     svgPanel.setAttribute('viewBox', '0 0 ' + String(width) + ' ' + String(height));
     svgPanel.setAttribute('width', width);
     svgPanel.setAttribute('height', height);
@@ -150,6 +118,7 @@ create.onclick = function () {
     svgPanelCoords = getCoords(svgPanel);
     updateRulers();
     updateGrid();
+    createFirstLayer();
 }
 
 //OPEN
@@ -171,20 +140,27 @@ function readFile(object) {
         clone.setAttribute("id", "svg_grid");
         let divs = document.querySelector("#draw_panel");
         let first = divs.firstElementChild;
+        first.setAttribute("id", "svg_panel");
         svgPanel = document.getElementById(first.id);
         svgPanel.insertBefore(clone, svgPanel.firstElementChild);
-        svgGrid = document.getElementById("svg_grid");
-        svgBackground = document.getElementById("svg_background");
-        if (first.getAttribute('width') || first.getAttribute('height') == null) {
+        let width = first.getAttribute('width');
+        let height = first.getAttribute('height')
+        if (width == null || height == null) {
             first.setAttribute('width', 512);
             first.setAttribute('height', 512);
             first.setAttribute('viewBox', '0 0 512 512');
         }
-        first.setAttribute('style', 'background-color: #fff; position: absolute; left: 50%; top: 50%; transform: translate(-50%, -50%); z-index: 1;');
+
+        svgPanel = document.getElementById(first.id);
+        svgPanel.setAttribute('style', 'background-color: #fff; position: absolute;');
+        centralLocation(width, height);
         scrollcoords = getCoords(scrollPanel);
         svgPanelCoords = getCoords(svgPanel);
         updateRulers();
         updateGrid();
+        for (var i = 0; i < layersPanel.childNodes.length;)
+            layersPanel.removeChild(layersPanel.childNodes[i]);
+        createFirstLayer();
     };
     reader.readAsText(file);
 }
@@ -199,15 +175,15 @@ save.onclick = function () {
     if (isGridEnabled) {
         flag = true;
         showGrid.click();
-    } 
+    }
     let fileName = prompt('Введите имя файла без расширения:');
     if (fileName == null) {
-       if (flag)
-            showGrid.click(); 
+        if (flag)
+            showGrid.click();
         return;
     }
     if (currentObject != null) {
-        currentObject.removeFrameAndPoints();
+        currentObject.hideFrameAndPoints();
         currentObject = null;
     }
     let svgData = draw_panel.innerHTML.toString();
@@ -223,7 +199,7 @@ save.onclick = function () {
     a.click();
 
     window.URL.revokeObjectURL(url);
-    if (flag) 
+    if (flag)
         showGrid.click();
 }
 
@@ -235,15 +211,15 @@ savePng.onclick = function () {
     if (isGridEnabled) {
         flag = true;
         showGrid.click();
-    } 
+    }
     let fileName = prompt('Введите имя файла без расширения:');
     if (fileName == null) {
-       if (flag)
-            showGrid.click(); 
+        if (flag)
+            showGrid.click();
         return;
     }
     if (currentObject != null) {
-        currentObject.removeFrameAndPoints();
+        currentObject.hideFrameAndPoints();
         currentObject = null;
     }
     let svgData = draw_panel.innerHTML.toString();
@@ -266,7 +242,7 @@ savePng.onclick = function () {
         window.URL.revokeObjectURL(url);
     }
     img.src = url;
-    if (flag) 
+    if (flag)
         showGrid.click();
 }
 
@@ -276,7 +252,7 @@ zoomIn.onclick = function () {
     svgPanelCoords = getCoords(svgPanel);
     svgPanel.style.transform = "translate(0, 0)";
     svgPanel.style.left = svgPanelCoords.left - scrollcoords.left;
-    svgPanel.style.top= svgPanelCoords.top - scrollcoords.top;
+    svgPanel.style.top = svgPanelCoords.top - scrollcoords.top;
     svgPanel.style.width = svgPanel.clientWidth * 1.5 + "px";
     svgPanel.style.height = svgPanel.clientHeight * 1.5 + "px";
     scaleСoef *= 1.5;
@@ -288,7 +264,7 @@ zoomOut.onclick = function () {
     svgPanelCoords = getCoords(svgPanel);
     svgPanel.style.transform = "translate(0, 0)";
     svgPanel.style.left = svgPanelCoords.left - scrollcoords.left;
-    svgPanel.style.top= svgPanelCoords.top - scrollcoords.top;
+    svgPanel.style.top = svgPanelCoords.top - scrollcoords.top;
     svgPanel.style.width = svgPanel.clientWidth / 1.5 + "px";
     svgPanel.style.height = svgPanel.clientHeight / 1.5 + "px";
     scaleСoef /= 1.5;
@@ -300,9 +276,9 @@ frontObject = document.getElementById("frontObject");
 
 frontObject.onclick = function () {
     if (currentObject != null) {
-        svgPanel.append(currentObject.svgElement);
-        for (let i = 0; i < currentObject.frame.length; i++) {
-            svgPanel.append(currentObject.frame[i].svgElement);
+        currentLayer.group.append(currentObject.svgElement);
+        for (let i = 0; i < currentObject.frameArray.length; i++) {
+            svgPanel.append(currentObject.frameArray[i].svgElement);
         }
         for (let i = 0; i < currentObject.pointsArray.length; i++) {
             svgPanel.append(currentObject.pointsArray[i].circle);
@@ -314,7 +290,7 @@ backObject = document.getElementById("backObject");
 
 backObject.onclick = function () {
     if (currentObject != null) {
-        svgPanel.prepend(currentObject.svgElement);
+        currentLayer.group.prepend(currentObject.svgElement);
     }
 }
 
@@ -330,11 +306,11 @@ showRulers.onclick = function () {
 }
 
 //SHOW GRID 
-showGrid = document.getElementById("showGrid"); 
+showGrid = document.getElementById("showGrid");
 showGrid.onclick = function () {
     if (!isGridEnabled) {
         isGridEnabled = true;
-        svgBackground.setAttribute("fill", "url(#grid_pattern)"); 
+        svgBackground.setAttribute("fill", "url(#grid_pattern)");
     } else {
         isGridEnabled = false;
         svgBackground.setAttribute("fill", "rgb(255, 255, 255)");
